chore: format script and add readme
